Move argparse-related things to new file
To reduce clutter in the original file.
