+++ augur/dates/__init__.py
@@ -1,6 +1,4 @@
-import argparse
 import datetime
-from textwrap import dedent
 import isodate
 import pandas as pd
 import re
@@ -9,12 +7,6 @@ from augur.errors import AugurError
 from .errors import InvalidDate, InvalidDateMessage
 
 from .ambiguous_date import AmbiguousDate
-
-SUPPORTED_DATE_HELP_TEXT = dedent("""\
-    1. an Augur-style numeric date with the year as the integer part (e.g. 2020.42) or
-    2. a date in ISO 8601 date format (i.e. YYYY-MM-DD) (e.g. '2020-06-04') or
-    3. a backwards-looking relative date in ISO 8601 duration format with optional P prefix (e.g. '1W', 'P1W')
-""")
 
 
 # Matches floats (e.g. 2018.0, -2018.0).
@@ -111,7 +103,7 @@ def _numeric_date(date, fmt=None, min_max_year=None, ambiguity_resolver=None):
     except (ValueError, isodate.ISO8601Error):
         pass
 
-    raise InvalidDateMessage(f"Ensure it is in one of the supported formats:\n{SUPPORTED_DATE_HELP_TEXT}")
+    raise InvalidDateMessage("Ensure it is in one of the supported formats.")
 
 
 def custom_strptime(date, fmt):
@@ -153,42 +145,6 @@ def relative_iso_to_datetime_date(backwards_duration_str: str, from_date: dateti
     if not backwards_duration_str.startswith('P'):
         backwards_duration_str = 'P'+backwards_duration_str
     return from_date - isodate.parse_duration(backwards_duration_str)
-
-
-def numeric_date_type_min(date):
-    """Get the numeric date from any supported date format, taking the minimum possible value if ambiguous.
-
-    This function is intended to be used as the `type` parameter in `argparse.ArgumentParser.add_argument()`
-
-    This raises an ArgumentTypeError from InvalidDate exceptions, otherwise the custom exception message won't be shown in console output due to:
-    https://github.com/python/cpython/blob/5c4d1f6e0e192653560ae2941a6677fbf4fbd1f2/Lib/argparse.py#L2503-L2513
-
-    >>> round(numeric_date_type_min("2018"), 3)
-    2018.001
-    """
-    try:
-        # TODO: support custom formats
-        return numeric_date(date, fmt="%Y-%m-%d", ambiguity_resolver='min')
-    except InvalidDate as e:
-        raise argparse.ArgumentTypeError(str(e)) from e
-
-
-def numeric_date_type_max(date):
-    """Get the numeric date from any supported date format, taking the maximum possible value if ambiguous.
-
-    This function is intended to be used as the `type` parameter in `argparse.ArgumentParser.add_argument()`
-
-    This raises an ArgumentTypeError from InvalidDate exceptions, otherwise the custom exception message won't be shown in console output due to:
-    https://github.com/python/cpython/blob/5c4d1f6e0e192653560ae2941a6677fbf4fbd1f2/Lib/argparse.py#L2503-L2513
-
-    >>> round(numeric_date_type_max("2018"), 3)
-    2018.999
-    """
-    try:
-        # TODO: support custom formats
-        return numeric_date(date, fmt="%Y-%m-%d", ambiguity_resolver='max')
-    except InvalidDate as e:
-        raise argparse.ArgumentTypeError(str(e)) from e
 
 
 def is_date_ambiguous(date, ambiguous_by="any"):
