+++ augur/frequencies.py
@@ -9,7 +9,8 @@ from Bio.Align import MultipleSeqAlignment
 
 from .frequency_estimators import get_pivots, alignment_frequencies, tree_frequencies
 from .frequency_estimators import AlignmentKdeFrequencies, TreeKdeFrequencies, TreeKdeFrequenciesError
-from .dates import numeric_date_type_min, numeric_date_type_max, SUPPORTED_DATE_HELP_TEXT, get_numerical_dates
+from .dates import get_numerical_dates
+from .dates.argparse import numeric_date_type_min, numeric_date_type_max, SUPPORTED_DATE_HELP_TEXT
 from .io.metadata import read_metadata
 from .utils import read_node_data, write_json
