+++ docs/api/developer/augur.dates.argparse.rst
@@ -1,0 +1,7 @@
+augur.dates.argparse
+====================
+
+.. automodule:: augur.dates.argparse
+   :members:
+   :undoc-members:
+   :show-inheritance:
